requirements.txt
django==1.8.6
django-secure==1.0.1
psycopg2==2.6.1
djangorestframework==3.3.0
django-cors-headers==1.1.0
django-heroku-memcacheify==0.8
django-suit==0.3a1
boto==2.38.0
django-environ==0.4.0
gunicorn==19.3.0
django-storages==1.1.8
gevent==1.1rc1
